requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
openpyxl==3.1.2
python-dotenv==1.0.0
alembic==1.11.3
